tests/wpt/tests/css/css-text/parsing/word-space-transform-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing word-space-transform with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-text-4/#word-space-transform">
<meta name="assert" content="word-space-transform supports only the grammar 'none | [ space | ideographic-space ] && auto-phrase?'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("word-space-transform", "auto");
test_invalid_value("word-space-transform", "normal");
test_invalid_value("word-space-transform", "spaces");
test_invalid_value("word-space-transform", "space ideographic-space");
test_invalid_value("word-space-transform", "auto-phrase");
test_invalid_value("word-space-transform", "none auto-phrase");
</script>
</body>
</html>
